pyproject.toml
[project]
name = "rag-lite"
version = "0.1.0"
description = "A lightweight RAG (Retrieval-Augmented Generation) system built with Flask"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "flask-cors>=6.0.2",
    "sqlalchemy>=2.0.45",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-chroma>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "langchain-deepseek>=1.0.1",
    "langchain-ollama>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "minio>=7.2.20",
    "langchain-milvus>=0.3.3",
    "python-docx>=1.1.0",
    "pymupdf>=1.24.0",
    "pypdf>=6.6.0",
    "sentence-transformers>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-flask>=1.3.0",
    "pytest-mock>=3.12.0",
    "reportlab>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*", "static*", "templates*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
